Implement/fix constant pool writing for classes, packages, modules and classloaders (#17)
    * Implement/fix constant pool writing for classes, packages, modules and classloaders

    * Fetch JFR type IDs from host JDK instead of hardwiring them.

    * Precompute and cache type IDs

    * Fold writing of packages, modules and classloaders into JfrTypeRepository

    * Removed debug printlns

    * Renamed count variables to poolCount

    * Simplify some collections accessors

    * Return int from JfrTraceIdLoadBarrier.classCount()

    * Improve comment

    * Improve JfrTypes enum construction

    * Add previousEpoch field to all JfrSymbolRepository.getSymbolId()

    * Remove redundant cast

+++ dataset/GitHub_Java/oracle.graal/substratevm/src/com.oracle.svm.jfr/src/com/oracle/svm/jfr/SubstrateJVM.java
@@ -422,7 +422,7 @@ class SubstrateJVM {
 
     @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     public boolean isEnabled(JfrEvents event) {
-        return eventSettings[(int) event.getId()].isEnabled();
+        return eventSettings[event.getId()].isEnabled();
     }
 
     /** See {@link JVM#setThreshold}. */
